Started conversion to unsigned long long

+++ include/globals.h
@@ -2,9 +2,9 @@
 #define GLOBALS_H
 #include <allegro.h>
 
-extern long money;
-extern int money_per_click;
-extern int money_per_second;
+extern unsigned long long money;
+extern unsigned long long money_per_click;
+extern unsigned long long money_per_second;
 
 bool location_clicked(int min_x,int max_x,int min_y,int max_y);
